\name{randletter-package}
\alias{randletter-package}
\alias{randletter}
\docType{package}
\title{
\packageTitle{randletter}
}
\description{
\packageDescription{randletter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{randletter}
\packageIndices{randletter}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{randletter}

Maintainer: \packageMaintainer{randletter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
